Reduce dependencies between schema / attributeValue transforming
