+++ src/main/scala/org/embulk/input/dynamodb/item/DynamodbItemReader.scala
@@ -10,9 +10,6 @@ case class DynamodbItemReader(
 ) {
   private var currentItem: Map[String, DynamodbAttributeValue] = _
 
-  val converter: DynamodbAttributeValueEmbulkConverter =
-    DynamodbAttributeValueEmbulkConverter(schema)
-
   def nextItem: Boolean = {
     ite.nextOption() match {
       case Some(v) =>
@@ -24,33 +21,44 @@ case class DynamodbItemReader(
 
   def getSchema: DynamodbItemSchema = schema
 
+  def getTransformable(
+      name: String,
+      value: DynamodbAttributeValue
+  ): DynamodbAttributeValueEmbulkTypeTransformable = {
+    DynamodbAttributeValueEmbulkTypeTransformable(
+      value,
+      typeEnforcer = schema.getAttributeType(name),
+      timestampParser = schema.getTimestampParser(name)
+    )
+  }
+
   def getBoolean(column: Column): Option[Boolean] =
     currentItem
       .get(column.getName)
-      .flatMap(v => converter.withAttribute(column.getName, v).asBoolean)
+      .flatMap(v => getTransformable(column.getName, v).asBoolean)
 
   def getString(column: Column): Option[String] =
     currentItem
       .get(column.getName)
-      .flatMap(v => converter.withAttribute(column.getName, v).asString)
+      .flatMap(v => getTransformable(column.getName, v).asString)
 
   def getLong(column: Column): Option[Long] =
     currentItem
       .get(column.getName)
-      .flatMap(v => converter.withAttribute(column.getName, v).asLong)
+      .flatMap(v => getTransformable(column.getName, v).asLong)
 
   def getDouble(column: Column): Option[Double] =
     currentItem
       .get(column.getName)
-      .flatMap(v => converter.withAttribute(column.getName, v).asDouble)
+      .flatMap(v => getTransformable(column.getName, v).asDouble)
 
   def getTimestamp(column: Column): Option[Timestamp] =
     currentItem
       .get(column.getName)
-      .flatMap(v => converter.withAttribute(column.getName, v).asTimestamp)
+      .flatMap(v => getTransformable(column.getName, v).asTimestamp)
 
   def getJson(column: Column): Option[Value] =
     currentItem
       .get(column.getName)
-      .flatMap(v => converter.withAttribute(column.getName, v).asMessagePack)
+      .flatMap(v => getTransformable(column.getName, v).asMessagePack)
 }
